+++ WebServer.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 15.0.26730.15
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebServer", "WebServer\WebServer.csproj", "{BB942B0C-1DE5-4AB3-900B-58E97F3D1AB3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebServerFixtures", "WebServerFixtures\WebServerFixtures.csproj", "{129C4BAC-4D10-4745-A214-431A13DBAD79}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -17,10 +15,6 @@ Global
 		{BB942B0C-1DE5-4AB3-900B-58E97F3D1AB3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BB942B0C-1DE5-4AB3-900B-58E97F3D1AB3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BB942B0C-1DE5-4AB3-900B-58E97F3D1AB3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{129C4BAC-4D10-4745-A214-431A13DBAD79}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{129C4BAC-4D10-4745-A214-431A13DBAD79}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{129C4BAC-4D10-4745-A214-431A13DBAD79}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{129C4BAC-4D10-4745-A214-431A13DBAD79}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
