+++ WebServer/Apps/StaticApp/FileSystem.cs
@@ -29,6 +29,5 @@ namespace WebServer
                 return null;
             }
         }
-
     }
 }
