+++ WebServer/Apps/StaticApp/StaticApp.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.IO;
-using System.Net;
-
-namespace WebServer
+﻿namespace WebServer
 {
     public class StaticApp : IApp
     {
