--- WebServerFixtures/UnitTest1.cs
@@ -1,14 +1,0 @@
-using System;
-using Xunit;
-
-namespace WebServerFixtures
-{
-    public class RoutesFixtures
-    {
-        [Fact]
-        public void Year_Should_Return_JSON()
-        {
-
-        }
-    }
-}
